Fix bug of atoi.

+++ lib/string.c
@@ -154,7 +154,7 @@ unsigned long atoi(char *in)
 	char *tmp = in;
 	int i = 0;
 	int ratio = 1;
-	int val = 0;
+	unsigned long val = 0;
 	unsigned long ret = 0;
 	int nbit = 0, hex_xbit = 0;
 	char type = TYPE_DECIMAL;
